Add editor option for different data types
In the EphysSocketEditor, added options for multiple data types, such as signed and unsigned integers at 8, 16, and 32 bits, and floats at 32 and 64 bits.
Shifted existing configuration options to make room for this option.
Note that this commit only makes changes to the editor; the underlying code does not make use of these changes yet.